dom/indexedDB/test/test_file_os_delete.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    const READ_WRITE = "readwrite";

    const name = window.location.pathname;

    const objectStoreName = "Blobs";

    getUsage(grabFileUsageAndContinueHandler);
    let startUsage = yield undefined;

    const fileData1 = {
      key: 1,
      obj: { id: 1, file: getRandomFile("random.bin", 100000) }
    };
    const fileData2 = {
      key: 2,
      obj: { id: 1, file: getRandomFile("random.bin", 100000) }
    };

    {
      let request = indexedDB.open(name, 1);
      request.onerror = errorHandler;
      request.onupgradeneeded = grabEventAndContinueHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield undefined;

      is(event.type, "upgradeneeded", "Got correct event type");

      let db = event.target.result;
      db.onerror = errorHandler;

      let objectStore = db.createObjectStore(objectStoreName, { });

      objectStore.createIndex("index", "id", { unique: true });

      objectStore.add(fileData1.obj, fileData1.key);

      request = objectStore.add(fileData2.obj, fileData2.key);
      request.addEventListener("error", new ExpectError("ConstraintError", true));
      request.onsuccess = unexpectedSuccessHandler;
      yield undefined;

      event = yield undefined;

      is(event.type, "success", "Got correct event type");

      getUsage(grabFileUsageAndContinueHandler);
      let usage = yield undefined;

      is(usage, startUsage + fileData1.obj.file.size + fileData2.obj.file.size,
         "Correct file usage");

      let trans = db.transaction([objectStoreName], READ_WRITE);
      trans.objectStore(objectStoreName).delete(fileData1.key);
      trans.oncomplete = grabEventAndContinueHandler;
      event = yield undefined;

      is(event.type, "complete", "Got correct event type");

      getUsage(grabFileUsageAndContinueHandler);
      usage = yield undefined;

      is(usage, startUsage + fileData1.obj.file.size + fileData2.obj.file.size,
         "OS files exists");

      fileData1.obj.file = null;
      fileData2.obj.file = null;
    }

    scheduleGC();
    yield undefined;

    // This isn't really necessary but in order to ensure that our files have
    // been deleted we need to round-trip with the PBackground thread...
    let request = indexedDB.deleteDatabase(name + "this can't exist");
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    yield undefined;

    getUsage(grabFileUsageAndContinueHandler);
    let endUsage = yield undefined;

    is(endUsage, startUsage, "OS files deleted");

    finishTest();
    yield undefined;
  }
  </script>
  <script type="text/javascript;version=1.7" src="file.js"></script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
